Changed the "localhost" to a variable called "namedserver"  it is initialized in both the dbase.js and the serve.js files.
This keeps the declaration centralized ("sort of") so that the app can be potentially hosted on a server other than local host.  (Note:  I think the API keys would have to be altered for that kind of purpose.  I had some hiccups trying to get ti to work on my server.  This helped)